def the functions
